Split the GUI plugin tools into their own file
