+++ src/components/scenes/GuiPluginViewScene.js
@@ -7,11 +7,12 @@ import { WebView } from 'react-native-webview'
 import { Bridge, onMethod } from 'yaob'
 
 import { EdgeProvider } from '../../modules/UI/scenes/Plugins/EdgeProvider.js'
-import { type GuiPlugin, makePluginUri } from '../../types/GuiPluginTypes.js'
+import { type GuiPlugin } from '../../types/GuiPluginTypes.js'
 import { connect } from '../../types/reactRedux.js'
 import { type Dispatch, type RootState } from '../../types/reduxTypes.js'
 import { type RouteProp } from '../../types/routerTypes.js'
 import { javascript } from '../../util/bridge/injectThisInWebView.js'
+import { makePluginUri } from '../../util/GuiPluginTools.js'
 import { bestOfPlugins } from '../../util/ReferralHelpers.js'
 import { SceneWrapper } from '../common/SceneWrapper.js'
 import { handlePluginBack, setPluginScene } from '../navigation/GuiPluginBackButton.js'
